# docs/openapi/users.yaml
openapi: 3.0.3
info:
  title: users-service API
  version: 1.0.0
  description: API de gestión de usuarios y autenticación
servers:
  - url: http://localhost:8081
    description: Local
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión y obtener JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/users/owner:
    post:
      tags: [Users]
      summary: Crear usuario propietario (solo ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/users/employee:
    post:
      tags: [Users]
      summary: Crear usuario empleado (solo OWNER)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEmployeeRequest'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/users/customer:
    post:
      tags: [Users]
      summary: Create customer account (public)
      description: Self-registration endpoint for customers. No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/usuarios/{id}:
    get:
      tags: [Usuarios]
      summary: Obtener usuario por id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '404': { description: No encontrado }
  /api/v1/usuarios/{id}/activo:
    get:
      tags: [Validaciones]
      summary: Validar si el usuario está activo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
  /api/v1/users/{userId}/restaurant/{restaurantId}/is-employee:
    get:
      tags: [Validations]
      summary: Validate if user is employee of specific restaurant
      description: |
        Critical endpoint for HU-012 security validation.
        Validates if a given user ID has EMPLOYEE role and is assigned to the specified restaurant.
        Used by Orders Service to enforce restaurant-based access control.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          description: ID of the user to validate
          schema:
            type: integer
            format: int64
        - in: path
          name: restaurantId
          required: true
          description: ID of the restaurant to check assignment
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: boolean
                description: true if user is employee of the restaurant, false otherwise
        '400':
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRequest:
      type: object
      required: [firstName, lastName, document, phone, birthDate, email, password]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        document: { type: string }
        phone: { type: string, maxLength: 13 }
        birthDate: { type: string, format: date }
        email: { type: string, format: email }
        password: { type: string, format: password }
    UserEmployeeRequest:
      allOf:
        - $ref: '#/components/schemas/UserRequest'
      properties:
        restaurantId:
          type: integer
          format: int64
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        role: { type: string, enum: [ADMIN, OWNER, EMPLOYEE, CUSTOMER] }
        active: { type: boolean }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        userId: { type: integer, format: int64 }
        role: { type: string }
        expiresIn: { type: integer }
    UsuarioResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        nombre: { type: string }
        apellido: { type: string }
        correo: { type: string }
        rol: { type: string, enum: [ADMINISTRADOR, PROPIETARIO, EMPLEADO, CLIENTE] }
        activo: { type: boolean }
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when error occurred
        status:
          type: integer
          description: HTTP status code